Statsbot-main/requirements.txt
flask>=3.0.0
pyyaml>=6.0
bcrypt>=4.0.0
